Remove duplicate file, move Value.tsx into component and download translations
