+++ src/client/pages/CheckoutDetails/QuoteDetails/QuoteDetails.tsx
@@ -1,50 +1,48 @@
 import React from 'react'
 import { colorsV3 } from '@hedviginsurance/brand'
 import styled from '@emotion/styled'
+import { getQuoteDetails } from 'utils/getQuoteDetails'
 import { useTextKeys } from 'utils/textKeys'
-import { TextButton } from 'components/buttons'
-import { Row } from '../components/Row'
-import { Label } from '../components/Label'
-import { Value } from '../components/Value'
+import { OfferQuote } from '../../OfferNew/types'
 import { SubSection } from '../components/SubSection'
-import { Group } from '../components/Group'
 import { HorizontalSpacer } from '../components/HorizontalSpacer'
-import { HomeInsuranceTypeOfContract } from '../../OfferNew/utils'
 
-const ButtonWrapper = styled.div`
-  width: 100%;
+const { gray500, gray300 } = colorsV3
+
+const Row = styled.div`
+  font-size: 1rem;
   display: flex;
-  justify-content: center;
-  align-items: center;
-  padding: 1rem 0 0.5rem;
+  flex-direction: row;
+  justify-content: space-between;
+  padding: 0.5rem 0;
+  line-height: 1;
 `
 
-type MainQuoteDetails = {
-  id: string
-  birthDate: string
-  street: string
-  city: string | null
-  zipCode: string
-  numberCoInsured: number
-  livingSpace: number
-  typeOfContract: HomeInsuranceTypeOfContract
-  isStudent: boolean
-}
+const Label = styled.div`
+  color: ${gray500};
+`
+const Value = styled.div`
+  color: ${colorsV3.gray700};
+  text-align: right;
+`
 
-type DetailsGroup = {
-  key: string
-  value: string
-  label: string
-}[]
+const Group = styled.div`
+  display: flex;
+  flex-direction: column;
+  padding: 0.5rem 0;
+  border-bottom: 1px solid ${gray300};
 
-type Props = { detailsGroups: DetailsGroup[] }
+  :last-of-type {
+    border-bottom: none;
+  }
+`
 
-export const QuoteDetails = ({ detailsGroups }: Props) => {
+export const QuoteDetails = ({ mainQuote }: { mainQuote: OfferQuote }) => {
   const textKeys = useTextKeys()
 
   return (
-    <SubSection headlineText={textKeys.CHECKOUT_YOUR_HOME_SECTION_TITLE()}>
-      {detailsGroups.map((group, index) => (
+    <SubSection headlineText={textKeys.CHECKOUT_QUOTE_DETAILS_TITLE()}>
+      {getQuoteDetails(mainQuote, textKeys).map((group, index) => (
         <Group key={index}>
           {group.map(({ key, value, label }) => (
             <Row key={key}>
@@ -55,11 +53,6 @@ export const QuoteDetails = ({ detailsGroups }: Props) => {
           ))}
         </Group>
       ))}
-      <ButtonWrapper>
-        <TextButton color={colorsV3.purple900}>
-          {textKeys.CHECKOUT_EDIT_INFORMATION_BUTTON()}
-        </TextButton>
-      </ButtonWrapper>
     </SubSection>
   )
 }
