--- src/client/pages/CheckoutDetails/components/Value.tsx
@@ -1,7 +1,0 @@
-import styled from '@emotion/styled'
-import { colorsV3 } from '@hedviginsurance/brand'
-
-export const Value = styled.div`
-  color: ${colorsV3.gray700};
-  text-align: right;
-`
