+++ src/translations/da_DK.json
@@ -346,5 +346,6 @@
   "SIDEBAR_REQUEST_CANCELLATION_INSURANCE_NAME_TOOLTIP": "Vi vil gerne hjælpe til med at opsige din nuværende {INSURANCE_NAME} for dig. Din nye forsikring hos Hedvig træder i kraft, samtidig med at den gamle ophører. Du modtager en mail med mere information.",
   "CHECKOUT_FOOTER_CANCELLATION_INFO": "Ingen bindingstid",
   "CHECKOUT_FOOTER_CONTINUE_TO_PAYMENT": "Fortsæt til betaling",
-  "CHECKOUT_FOOTER_COMPLETE_PURCHASE": "Gennemfør køb"
+  "CHECKOUT_FOOTER_COMPLETE_PURCHASE": "Gennemfør køb",
+  "CHECKOUT_QUOTE_DETAILS_TITLE": "Opsummering"
 }
