+++ src/translations/en_DK.json
@@ -346,5 +346,6 @@
   "SIDEBAR_REQUEST_CANCELLATION_INSURANCE_NAME_TOOLTIP": "We will gladly help you cancel your current {INSURANCE_NAME} for you. Your new insurance with Hedvig will be activated when your old expires. You’ll receive more info via email.",
   "CHECKOUT_FOOTER_CANCELLATION_INFO": "Cancel anytime",
   "CHECKOUT_FOOTER_CONTINUE_TO_PAYMENT": "Continue to payment",
-  "CHECKOUT_FOOTER_COMPLETE_PURCHASE": "Complete purchase"
+  "CHECKOUT_FOOTER_COMPLETE_PURCHASE": "Complete purchase",
+  "CHECKOUT_QUOTE_DETAILS_TITLE": "Summary"
 }
