+++ src/translations/fr.json
@@ -346,5 +346,6 @@
   "SIDEBAR_REQUEST_CANCELLATION_INSURANCE_NAME_TOOLTIP": "",
   "CHECKOUT_FOOTER_CANCELLATION_INFO": "",
   "CHECKOUT_FOOTER_CONTINUE_TO_PAYMENT": "",
-  "CHECKOUT_FOOTER_COMPLETE_PURCHASE": ""
+  "CHECKOUT_FOOTER_COMPLETE_PURCHASE": "",
+  "CHECKOUT_QUOTE_DETAILS_TITLE": ""
 }
