+++ src/translations/nb_NO.json
@@ -346,5 +346,6 @@
   "SIDEBAR_REQUEST_CANCELLATION_INSURANCE_NAME_TOOLTIP": "Vi hjelper deg gjerne med å avslutte din nåværende {INSURANCE_NAME} for deg. Din nye forsikring hos Hedvig aktiveres når den gamle utløper. Du vil få mer informasjon på mail.",
   "CHECKOUT_FOOTER_CANCELLATION_INFO": "Ingen bindingstid",
   "CHECKOUT_FOOTER_CONTINUE_TO_PAYMENT": "Fortsett til betaling",
-  "CHECKOUT_FOOTER_COMPLETE_PURCHASE": "Gjennomfør kjøp"
+  "CHECKOUT_FOOTER_COMPLETE_PURCHASE": "Gjennomfør kjøp",
+  "CHECKOUT_QUOTE_DETAILS_TITLE": "Sammendrag"
 }
